Add the fed2020 overlay.
The 2020 tax return forms are likely to be quite similar to the 2019
versions, with inflation-adjustments to constant values. To avoid
copy-paste-tweak with the forms, use an overlay: forms with differences
from the previous year can be rewritten, but unchanged ones should just
use the prior year's form.

Constant values are now associated with an overlay TaxReturn class, so
the inflation adjustments do not require new forms.